requirements.txt
# Web framework
fastapi>=0.110.0,<1.0.0

# AWS Lambda adapter for FastAPI/ASGI
mangum>=0.17.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Data validation (Pydantic v2 — required for ConfigDict, field_validator)
pydantic>=2.6.0,<3.0.0

# Environment variable loading
python-dotenv>=1.0.0

# Local development server (not needed in Lambda, only for local runs)
# Use plain uvicorn without C extensions to avoid cross-platform issues
uvicorn>=0.29.0
